backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
pyjwt==2.8.0
Flask-Migrate==3.1.0
